Creating Engine to manage the state of the application, including players and board.
This is a GenServer worker, that will be started by the Application supervisor.
When a game is initiated we create a new game with an empty board and we define x and o players. That is returned by the api endpoint.
Also getting rid off an annoying error when running the app regarding es2015 not supported by installing npm package babel-es2015.
Unit testing the GenServer is like testing any other module except for the fact that we need to start it in the test setup.
For the unit tests of the controller I have mocked the Engine.