When game has a result we display the winner or the correspondent game result. User has then the possibility of restarting the button.
This is achieved by creating a game result store that listens to every move and checks if the status is winner or draw. Once that happens, it notifies listeners.
Create on main component GameOver that listens to the former store and opens a react modal dialog (winner or draw) with the game result.
If user clicks the button, we reload the page.

ReactModal component is nice and easy to use but he has given me a really hard time mocking it. Also ReactModal wants to render too early in the tests, even before the dom is ready. Have tried really hard to set some timeouts.
At the end, as usual, the best solution is to hide away external dependencies of the app as far away of the domain logic as possible. As usual better testability means a better design.
The way I reload the game means I reload the app. The cons of this is that the full javascript bundle is reloaded in the app. To follow YAGNI and keep it simple in the mean time, doing like this seemed a good balance.