My kickass computer player was not as clever as I thought... Introducing Prevent fork.
Force fork knows how to create a fork but it didn't know how to block a future fork the opponent.
On the other hand, block fork was not clever enough to envision a future opponent's fork.
Now Prevent Fork finds what empty position to play that blocks a future fork by the opponent.