Refactoring logic to calculate a fork, created Board Cutter.
In order to calculate a Fork we need to find positions that belong two several triples containing two empty spaces.
Fork moves was too big and it felt that it was violating single responsibility.
Now I have refactored that logic into Board cutter. It could potentially be also used for other strategies.