Now the state of the Game in GameUI is encapsulated in one place, TicTacToe component. Setting state is more explicit.
I have deleted all the intermediate stores. Now TicTacToe listens to move and game start completed and set the new state.
All the other components use props pass down from TicTacToe.
This solution doesn't use Reflux that intensively and it is a bit more heavy on passing down props but the advantage is that the state is move explicit. That makes it easy to follow the unidirectional dataflow so that maintain this code will probably be easier.